Team meeting commit
Apart from sites, we talked 1 hour just about ActivityPub `Collection` and `attachment` –
the reason was that both can take supermuch space in the UI and can be nested.

While the idea for Collections is to have a pagination navigation on top and treat nested
Collections by having the same below like in <ul>/<li> but instead bullet with a down-right arrow [or down-left for RTL] …
So you can browse / flip through them in any context (row/card/column/fullgrid) …

Attachments was the question.
Differences to Collection
- we would like to have the previews in the pagination row (e.g. play Audio/Video) !
- should they “triggerPopup” for fullscreen context, just like image with gallery ?

For now we decided not to expand them by default but instead have a <details> with count in summary and unique
type-icons which act as filter-labels.
This is all we add here for today :)