Dynamic Select
For widgets like ActivityPub locales, show up to n options as single checkbox/radio-group and more as Select widget …

Please note that the AP-id caching-middleware was not used in this repo yet. Will be in the next commit alongside the locales middleware.
Team meeting on SAT is 1 hour later: 17UTC