Image v1
With this commit the Image, Images and ImageCaption modules are in the final stage.
It is missing the location (own widget, can expand to ESRI/OSM) and instrument property for now …

Give /Image an ActivityPub Object and enjoy a gallery which is working without JS, can be statically rendered, is strictly baselined and themable …
And switchable between column/row and dark/light …
This is the “Object” part of a final “card”, it might have the Activity and potential Collection navigation above and the Action-buttons and Comments/Replies below …
See /Image for first example.